Move AZC0005 ClientSettings test into AZC0005Tests
